+++ modules/common/src/main/java/io/jpom/script/BaseRunScript.java
@@ -27,8 +27,8 @@ import cn.hutool.core.date.DateUnit;
 import cn.hutool.core.date.DateUtil;
 import cn.hutool.core.io.FileUtil;
 import cn.hutool.core.io.IoUtil;
+import io.jpom.JpomApplication;
 import io.jpom.common.Const;
-import io.jpom.system.ConfigBean;
 import io.jpom.util.LogRecorder;
 
 import java.io.File;
@@ -86,7 +86,7 @@ public abstract class BaseRunScript implements AutoCloseable {
      * 清理 脚本文件执行缓存
      */
     public static void clearRunScript() {
-        String dataPath = ConfigBean.getInstance().getDataPath();
+        String dataPath = JpomApplication.getInstance().getDataPath();
         File scriptFile = FileUtil.file(dataPath, Const.SCRIPT_RUN_CACHE_DIRECTORY);
         if (!FileUtil.isDirectory(scriptFile)) {
             return;